adding load method to FullReturnModel
