+++ app/controllers/BaseController.scala
@@ -57,10 +57,4 @@ trait BaseController extends FrontendBaseController with I18nSupport with Enumer
       case Some(ans) => f(ans)
       case _ => Future.successful(InternalServerError(errorHandler.internalServerErrorTemplate))
     }
-
-  def getSessionData(key: String)(f: String => Future[Result])(implicit request: Request[_], errorHandler: ErrorHandler): Future[Result] =
-    request.session.get(key) match {
-      case Some(data) => f(data)
-      case _ => Future.successful(InternalServerError(errorHandler.internalServerErrorTemplate))
-    }
 }
