+++ test/controllers/reviewAndComplete/ReviewAndCompleteControllerSpec.scala
@@ -21,21 +21,12 @@ import base.SpecBase
 import config.featureSwitch.FeatureSwitching
 import controllers.actions._
 import mocks.MockInterestRestrictionReturnConnector
-import models.FullOrAbbreviatedReturn.Full
-import models.{FullReturnModel, UserAnswers}
-import models.returnModels.{AccountingPeriodModel, CompanyNameModel, DeemedParentModel, ReviewAndCompleteModel, UTRModel}
+import models.FullReturnModel
+import models.returnModels.ReviewAndCompleteModel
 import navigation.FakeNavigators.FakeReviewAndCompleteNavigator
-import pages.aboutReturn.{AccountingPeriodPage, AgentActingOnBehalfOfCompanyPage, AgentNamePage, FullOrAbbreviatedReturnPage, ReportingCompanyAppointedPage, ReportingCompanyCTUTRPage, ReportingCompanyNamePage, TellUsWhatHasChangedPage}
-import pages.elections.{ElectedInterestAllowanceAlternativeCalcBeforePage, ElectedInterestAllowanceConsolidatedPshipBeforePage, GroupRatioElectionPage, InterestAllowanceConsolidatedPshipElectionPage, InterestAllowanceNonConsolidatedInvestmentsElectionPage}
-import pages.groupLevelInformation.{DisallowedAmountPage, EnterANGIEPage, GroupInterestAllowancePage, GroupInterestCapacityPage, GroupSubjectToRestrictionsPage, InterestAllowanceBroughtForwardPage, ReturnContainEstimatesPage, RevisingReturnPage}
 import pages.reviewAndComplete.ReviewAndCompletePage
-import pages.ultimateParentCompany.{DeemedParentPage, HasDeemedParentPage, ReportingCompanySameAsParentPage}
 import play.api.test.Helpers._
-import sectionstatus.{AboutReturnSectionStatus, ElectionsSectionStatus, GroupLevelInformationSectionStatus, UltimateParentCompanySectionStatus}
 import views.html.reviewAndComplete.ReviewAndCompleteView
-
-import java.time.LocalDate
-
 
 class ReviewAndCompleteControllerSpec extends SpecBase with FeatureSwitching with BaseConstants with MockDataRetrievalAction with MockInterestRestrictionReturnConnector {
 
@@ -66,54 +57,9 @@ class ReviewAndCompleteControllerSpec extends SpecBase with FeatureSwitching wit
     }
 
     "redirect to the next page when submitted" in {
-      val date = LocalDate.of(2020,1,1)
-      val deemedParentModel =
-        DeemedParentModel(
-          companyName = CompanyNameModel("Name"),
-          sautr = Some(UTRModel("1123456789")),
-          limitedLiabilityPartnership = Some(true),
-          payTaxInUk = Some(true),
-          reportingCompanySameAsParent = Some(false)
-        )
-
-      val userAnswers = UserAnswers("id")
-        .set(ReportingCompanyAppointedPage, true).get
-        .set(AgentActingOnBehalfOfCompanyPage, true).get
-        .set(AgentNamePage, agentName).get
-        .set(FullOrAbbreviatedReturnPage, Full).get
-        .set(RevisingReturnPage, true).get
-        .set(TellUsWhatHasChangedPage, "Something has changed").get
-        .set(ReportingCompanyNamePage, companyNameModel.name).get
-        .set(ReportingCompanyCTUTRPage, ctutrModel.utr).get
-        .set(AccountingPeriodPage, AccountingPeriodModel(date, date.plusMonths(1))).get
-        .set(GroupRatioElectionPage, false).get
-        .set(GroupSubjectToRestrictionsPage, true).get
-        .set(DisallowedAmountPage, BigDecimal(123.12)).get
-        .set(InterestAllowanceBroughtForwardPage, BigDecimal(123.12)).get
-        .set(GroupInterestAllowancePage, BigDecimal(123.12)).get
-        .set(GroupInterestCapacityPage, BigDecimal(123.12)).get
-        .set(EnterANGIEPage, BigDecimal(123.12)).get
-        .set(ReturnContainEstimatesPage, false).get
-        .set(ReportingCompanySameAsParentPage, false).get
-        .set(HasDeemedParentPage, false).get
-        .set(DeemedParentPage, deemedParentModel, Some(1)).get
-        .set(GroupRatioElectionPage, false).get
-        .set(ElectedInterestAllowanceAlternativeCalcBeforePage,true).get
-        .set(InterestAllowanceNonConsolidatedInvestmentsElectionPage,false).get
-        .set(ElectedInterestAllowanceConsolidatedPshipBeforePage,false).get
-        .set(InterestAllowanceConsolidatedPshipElectionPage,false).get
-
-
-      val aboutReturnStatus = AboutReturnSectionStatus.currentSection(userAnswers)
-      val electionsStatus = ElectionsSectionStatus.currentSection(userAnswers)
-      val groupLevelInformationStatus = GroupLevelInformationSectionStatus.currentSection(userAnswers)
-      val ultimateParentCompanyStatus = UltimateParentCompanySectionStatus.currentSection(userAnswers)
-
-      val irr = FullReturnModel(aboutReturnStatus.get,ultimateParentCompanyStatus.get,electionsStatus.get,groupLevelInformationStatus,None)
-
-      mockGetAnswers(Some(userAnswers))
+      mockGetAnswers(Some(completeUserAnswers))
       mockSetAnswers()
-      mockSubmitReturn(irr)
+      mockSubmitReturn(FullReturnModel.load(completeUserAnswers).get)
 
       val result = Controller.onSubmit()(fakeDataRequest)
